+++ Plugin/src/MapStuff/BiomeManager.cs
@@ -3,6 +3,7 @@ using Unity.Netcode;
 using UnityEngine;
 using UnityEngine.Rendering.HighDefinition;
 using System.Collections;
+using System.Diagnostics;
 
 namespace CodeRebirth.MapStuff
 {
@@ -12,10 +13,11 @@ namespace CodeRebirth.MapStuff
         public DecalProjector crimsonProjector = null!;
         public DecalProjector hallowProjector = null!;
         
-        private ParticleSystem deadParticles = null!;
+        private ParticleSystem deathParticles = null!;
         private DecalProjector activeProjector = null!;
         private System.Random random = new System.Random(69);
         private int foliageLayer;
+        private int terrainLayer;
 
         public void Start()
         {
@@ -40,16 +42,17 @@ namespace CodeRebirth.MapStuff
                     break;
             }
 
-            deadParticles = activeProjector.gameObject.GetComponentInChildren<ParticleSystem>();
+            deathParticles = activeProjector.gameObject.GetComponentInChildren<ParticleSystem>();
             foliageLayer = LayerMask.NameToLayer("Foliage");
+            terrainLayer = LayerMask.NameToLayer("Terrain");
             StartCoroutine(CheckAndDestroyFoliage());
         }
 
         public void Update()
         {
             if (activeProjector == null) return;
-
-            activeProjector.gameObject.transform.localScale += new Vector3(Time.deltaTime * 0.34f, Time.deltaTime * 0.34f, 0f);
+            if (activeProjector.size.x >= 350 || activeProjector.size.y >= 350) return;
+            activeProjector.size += new Vector3(Time.deltaTime * 0.34f, Time.deltaTime * 0.34f, 0f);
         }
 
         private IEnumerator CheckAndDestroyFoliage()
@@ -62,38 +65,67 @@ namespace CodeRebirth.MapStuff
         }
 
         private void PerformSphereCast()
-        { // todo: make it destroy destroyable trees too.
-            Collider[] hitColliders = Physics.OverlapSphere(activeProjector.transform.position, activeProjector.gameObject.transform.localScale.y/4f, 1 << foliageLayer);
+        {
+            Stopwatch timer = new Stopwatch();
+            timer.Start();
+            // Combine the foliage layer with the terrain layer to perform sphere cast on both
+            int combinedLayerMask = (1 << foliageLayer) | (1 << terrainLayer);
+            
+            // Perform sphere cast
+            Collider[] hitColliders = Physics.OverlapSphere(activeProjector.transform.position, activeProjector.size.y / 4f, combinedLayerMask);
             foreach (var hitCollider in hitColliders)
             {
-                var meshRenderer = hitCollider.GetComponent<MeshRenderer>();
-                var meshFilter = hitCollider.GetComponent<MeshFilter>();
-
-                if (meshRenderer != null && meshFilter != null)
+                // Check if the collider belongs to foliage or a tree
+                if (IsFoliage(hitCollider) || IsTree(hitCollider))
                 {
-                    Vector3 hitPosition = hitCollider.transform.position;
-                    Quaternion hitRotation = hitCollider.transform.rotation;
-
-                    if (hitCollider.GetComponent<NetworkObject>() != null)
-                    {
-                        // Despawn if network spawned
-                        if (IsServer)
-                        {
-                            hitCollider.GetComponent<NetworkObject>().Despawn();
-                        }
-                    }
-                    else
-                    {
-                        // Destroy if locally spawned
-                        Destroy(hitCollider.gameObject);
-                    }
-
-                    // Instantiate dead particles at the position of the destroyed foliage
-                    ParticleSystem particles = Instantiate(deadParticles, hitPosition, hitRotation);
-                    particles.Play();
-                    Destroy(particles.gameObject, particles.main.duration + particles.main.startLifetime.constantMax);
+                    DestroyColliderObject(hitCollider);
                 }
             }
+
+            timer.Stop();
+            Plugin.ExtendedLogging($"Run completed in {timer.ElapsedTicks} ticks and {timer.ElapsedMilliseconds}ms");
+        }
+
+        private bool IsFoliage(Collider collider)
+        {
+            var meshRenderer = collider.GetComponent<MeshRenderer>();
+            var meshFilter = collider.GetComponent<MeshFilter>();
+            return meshRenderer != null && meshFilter != null && collider.gameObject.layer == foliageLayer;
+        }
+
+        private bool IsTree(Collider collider)
+        {
+            var meshRenderer = collider.GetComponent<MeshRenderer>();
+            var meshFilter = collider.GetComponent<MeshFilter>();
+            return meshRenderer != null && meshFilter != null &&
+                collider.CompareTag("Wood") &&
+                collider.gameObject.layer == terrainLayer &&
+                collider != null && !collider.isTrigger;
+        }
+
+        private void DestroyColliderObject(Collider collider)
+        {
+            Vector3 hitPosition = collider.transform.position;
+            Quaternion hitRotation = collider.transform.rotation;
+
+            if (collider.GetComponent<NetworkObject>() != null)
+            {
+                // Despawn if network spawned
+                if (IsServer)
+                {
+                    collider.GetComponent<NetworkObject>().Despawn();
+                }
+            }
+            else
+            {
+                // Destroy if locally spawned
+                Destroy(collider.gameObject);
+            }
+
+            // Instantiate dead particles at the position of the destroyed object
+            ParticleSystem particles = Instantiate(deathParticles, hitPosition, hitRotation);
+            particles.Play();
+            Destroy(particles.gameObject, particles.main.duration + particles.main.startLifetime.constantMax);
         }
     }
 }