+++ Plugin/Thunderstore/CHANGELOG.md
@@ -6,6 +6,8 @@
 - Improved some patches.
 - Added dangerous flora.
 - Added new companion, robot doggy.
+- Changed the weathers to have a config to not do the powerlevel shenanigans, but currently tornado and meteor shower increase inside power by 6 and decrease outside and daytime by 3.
+- Fixed money radar icon not disappearing when picked up.
 
 ## v0.4.2
 - Sigh, fixed the tornado stuff for the guy who wants it to spawn naturally as an enemy (WONT WORK IF THE WEATHER IS WIMDY ALREADY).